refactor- result
