+++ Assets/AudioRecorder/Scripts/Runtime/Utility/Result.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4d94dbc343ce8ea4db75d18132e37f39
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
