+++ Assets/AudioRecorder/Scripts/Runtime/Utility/Result/AudioClipFileReadingResultModel.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace Mayank.AudioRecorder.Utility.Result
+{
+    public class AudioClipFileReadingResultModel : ResultModel<AudioClip>
+    {
+        
+    }
+}
