+++ Assets/AudioRecorder/Scripts/Runtime/Utility/Result/AudioClipFileReadingResultModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f5060c2c225759d47b89a8597cf3a873
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
