+++ Assets/AudioRecorder/Scripts/Runtime/Utility/Result/FileWritingResultModel.cs
@@ -1,0 +1,21 @@
+namespace Mayank.AudioRecorder.Utility.Result
+{
+    // public class FileWritingResultModel
+    public class FileWritingResultModel : ResultModel<string>
+    {
+        // /// <summary>
+        // /// The status of the file writing operation.
+        // /// </summary>
+        // public bool status;
+        //
+        // /// <summary>
+        // /// The result of the file writing operation.
+        // /// </summary>
+        // public string result = null;
+        //
+        // /// <summary>
+        // /// The error message, if any, encountered during the file writing operation.
+        // /// </summary>
+        // public string error = null;
+    }
+}
