+++ Assets/AudioRecorder/Scripts/Runtime/Utility/Result/ResultModel.cs
@@ -1,6 +1,6 @@
-namespace Mayank.AudioRecorder.Utility
+namespace Mayank.AudioRecorder.Utility.Result
 {
-    public class FileWritingResultModel
+    public class ResultModel<T>
     {
         /// <summary>
         /// The status of the file writing operation.
@@ -10,7 +10,8 @@ namespace Mayank.AudioRecorder.Utility
         /// <summary>
         /// The result of the file writing operation.
         /// </summary>
-        public string result = null;
+        // public string result = null;
+        public T result;
         
         /// <summary>
         /// The error message, if any, encountered during the file writing operation.
