+++ Assets/AudioRecorder/Scripts/Runtime/Utility/Result/ResultModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f6426883042e4c449d865650b772df3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
